# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PropEdge API - sports betting stats tracker and community
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List players
      parameters:
        - name: sport
          in: query
          schema: { type: string }
        - name: teamId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}/stats/{period}:
    get:
      operationId: getPlayerStats
      tags: [players]
      summary: Get player game-by-game stats for a period
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: period
          in: path
          required: true
          schema:
            type: string
            enum: [week, biweek, month, season]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerGameStat"
  /players/{id}/prop-summary:
    get:
      operationId: getPlayerPropSummary
      tags: [players]
      summary: Averaged prop stats for a player (useful for betting lines)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerPropSummary"

  # ── Teams ────────────────────────────────────────────────────────────────
...
  /bets:
    get:
      operationId: listBets
      tags: [bets]
      summary: List bets for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, won, lost, push, void]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bet"
    post:
      operationId: createBet
      tags: [bets]
      summary: Log a new bet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
...
  /bets/{id}:
    patch:
      operationId: updateBet
      tags: [bets]
      summary: Update a bet (settle result)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
    delete:
      operationId: deleteBet
      tags: [bets]
      summary: Delete a bet
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Simulator ────────────────────────────────────────────────────────────
...
  /simulator/bets:
    get:
      operationId: listSimulatorBets
      tags: [simulator]
      summary: List virtual bets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulatorBet"
    post:
      operationId: createSimulatorBet
      tags: [simulator]
      summary: Place a virtual bet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulatorBetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorBet"
  /simulator/bets/{id}:
    patch:
      operationId: settleSimulatorBet
      tags: [simulator]
      summary: Settle a virtual bet (win/loss)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulatorBetSettle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorBet"

  # ── Community ────────────────────────────────────────────────────────────
  /posts:
    get:
      operationId: listPosts
      tags: [community]
      summary: Get community feed
      parameters:
        - name: cursor
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostFeed"
    post:
      operationId: createPost
      tags: [community]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [community]
      summary: Get a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [community]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [community]
      summary: Like or unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"

  # ── Users ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, name, sport, position, teamId, teamName]
      properties:
        id:        { type: integer }
        name:      { type: string }
        sport:     { type: string }
        position:  { type: string }
        teamId:    { type: integer }
        teamName:  { type: string }
        number:    { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
    PlayerGameStat:
      type: object
      required: [id, playerId, gameDate, opponent]
      properties:
        id:           { type: integer }
        playerId:     { type: integer }
        gameDate:     { type: string }
        opponent:     { type: string }
        # Basketball / shared
        points:       { type: ["number", "null"] }
        rebounds:     { type: ["number", "null"] }
        assists:       { type: ["number", "null"] }
        steals:       { type: ["number", "null"] }
        blocks:       { type: ["number", "null"] }
        threePointers: { type: ["number", "null"] }
        turnovers:    { type: ["number", "null"] }
        minutesPlayed: { type: ["number", "null"] }
        # Baseball
        hits:         { type: ["number", "null"] }
        homeRuns:     { type: ["number", "null"] }
        rbis:         { type: ["number", "null"] }
        runs:         { type: ["number", "null"] }
        strikeouts:   { type: ["number", "null"] }
        walks:        { type: ["number", "null"] }
        # Pitcher stats
        inningsPitched: { type: ["number", "null"] }
        earnedRuns:   { type: ["number", "null"] }
    PlayerPropSummary:
      type: object
      required: [playerId, playerName, sport, gamesPlayed]
      properties:
        playerId:      { type: integer }
        playerName:    { type: string }
        sport:         { type: string }
        gamesPlayed:   { type: integer }
        avgPoints:     { type: ["number", "null"] }
        avgRebounds:   { type: ["number", "null"] }
        avgAssists:    { type: ["number", "null"] }
        avgSteals:     { type: ["number", "null"] }
        avgThreePointers: { type: ["number", "null"] }
        avgHits:       { type: ["number", "null"] }
        avgHomeRuns:   { type: ["number", "null"] }
        avgRbis:       { type: ["number", "null"] }
...
    Bet:
      type: object
      required: [id, userId, description, betType, wager, odds, status, createdAt]
      properties:
        id:           { type: integer }
        userId:       { type: integer }
        description:  { type: string }
        betType:      { type: string }
        sportsbook:   { type: ["string", "null"] }
        wager:        { type: number }
        odds:         { type: number }
        potentialPayout: { type: number }
        actualPayout: { type: ["number", "null"] }
        status:       { type: string, enum: [pending, won, lost, push, void] }
        sport:        { type: ["string", "null"] }
        playerName:   { type: ["string", "null"] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
        settledAt:    { type: ["string", "null"] }
    BetInput:
      type: object
      required: [description, betType, wager, odds]
      properties:
        description:  { type: string }
        betType:      { type: string }
        sportsbook:   { type: string }
        wager:        { type: number }
        odds:         { type: number }
        sport:        { type: string }
        playerName:   { type: string }
        notes:        { type: string }
    BetUpdate:
      type: object
      properties:
        status:       { type: string, enum: [pending, won, lost, push, void] }
        actualPayout: { type: number }
        notes:        { type: string }
...
    SimulatorBet:
      type: object
      required: [id, userId, description, betType, wager, odds, potentialPayout, status, createdAt]
      properties:
        id:              { type: integer }
        userId:          { type: integer }
        description:     { type: string }
        betType:         { type: string }
        wager:           { type: number }
        odds:            { type: number }
        potentialPayout: { type: number }
        actualPayout:    { type: ["number", "null"] }
        status:          { type: string, enum: [pending, won, lost, push] }
        sport:           { type: ["string", "null"] }
        playerName:      { type: ["string", "null"] }
        createdAt:       { type: string }
    SimulatorBetInput:
      type: object
      required: [description, betType, wager, odds]
      properties:
        description: { type: string }
        betType:     { type: string }
        wager:       { type: number }
        odds:        { type: number }
        sport:       { type: string }
        playerName:  { type: string }
    SimulatorBetSettle:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [won, lost, push] }

    # ── Community ────────────────────────────────────────────────────────
    Post:
      type: object
      required: [id, userId, username, content, likeCount, liked, createdAt]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        username:    { type: string }
        avatarUrl:   { type: ["string", "null"] }
        content:     { type: string }
        likeCount:   { type: integer }
        liked:       { type: boolean }
        sport:       { type: ["string", "null"] }
        playerTag:   { type: ["string", "null"] }
        createdAt:   { type: string }
    PostFeed:
      type: object
      required: [posts, hasMore]
      properties:
        posts:   { type: array, items: { $ref: "#/components/schemas/Post" } }
        hasMore: { type: boolean }
        nextCursor: { type: ["integer", "null"] }
    PostInput:
      type: object
      required: [content]
      properties:
        content:   { type: string }
        sport:     { type: string }
        playerTag: { type: string }
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:     { type: boolean }
        likeCount: { type: integer }

    # ── Users ────────────────────────────────────────────────────────────
...