# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PropEdge API - sports betting stats tracker and community
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  sport: string;
  position: string;
  teamId: number;
  teamName: string;
  /** @nullable */
  number?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}