# lib/api-zod/src/generated/types/playerPropSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PropEdge API - sports betting stats tracker and community
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerPropSummary {
  playerId: number;
  playerName: string;
  sport: string;
  gamesPlayed: number;
  /** @nullable */
  avgPoints?: number | null;
  /** @nullable */
  avgRebounds?: number | null;
  /** @nullable */
  avgAssists?: number | null;
  /** @nullable */
  avgSteals?: number | null;
  /** @nullable */
  avgThreePointers?: number | null;
  /** @nullable */
  avgHits?: number | null;
  /** @nullable */
  avgHomeRuns?: number | null;
  /** @nullable */
  avgRbis?: number | null;
}