# lib/api-zod/src/generated/types/simulatorBet.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PropEdge API - sports betting stats tracker and community
 * OpenAPI spec version: 0.1.0
 */
import type { SimulatorBetStatus } from './simulatorBetStatus';

export interface SimulatorBet {
  id: number;
  userId: number;
  description: string;
  betType: string;
  wager: number;
  odds: number;
  potentialPayout: number;
  /** @nullable */
  actualPayout?: number | null;
  status: SimulatorBetStatus;
  /** @nullable */
  sport?: string | null;
  /** @nullable */
  playerName?: string | null;
  createdAt: string;
}