# lib/api-zod/src/generated/types/simulatorBetInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PropEdge API - sports betting stats tracker and community
 * OpenAPI spec version: 0.1.0
 */

export interface SimulatorBetInput {
  description: string;
  betType: string;
  wager: number;
  odds: number;
  sport?: string;
  playerName?: string;
}